pyproject.toml
[project]
name = "melodai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "pycryptodome>=3.21.0",
    "python-dotenv>=1.0.1",
    "replicate>=1.0.3",
    "requests>=2.32.3",
    "flask-socketio>=5.4.1",
    "groq>=0.12.0",
]
